// src/component/comparisonItem/schemas/OldCollege.ts
export class OldCollege {
    id: string;
    name?: string;
    image?: string;
    city?: string;
    state?: string;
    description?: string;
    founded_date?: string;
    website?: string;
    status?: boolean;
    rating: number;
    image_author?: string;
    image_title?: string;
    image_url?: string;
    image_license?: string;
}